Move api calls for application form into api folder
